Updated ES5 to ES6
