+++ backend/fetchContests.js
@@ -1,10 +1,12 @@
-const https = require('https');
-require('dotenv').config();
+import https from "https";
+import dotenv from 'dotenv';
+dotenv.config();
+
 async function fetchContestsData() {
   return new Promise((resolve, reject) => {
     const options = {
       hostname: process.env.BACKEND_URL,
-      path: '/contests',
+      path: '/api/contests',
       method: 'GET',
     };
 
@@ -28,4 +30,4 @@ async function fetchContestsData() {
   });
 }
 
-module.exports = fetchContestsData;
+export default fetchContestsData;
