--- backend/users/services/getUser.js
@@ -1,17 +1,0 @@
-const User = require('../models/User');
-
-const getUser = async (identifier) => {
-    try {
-        const user = await User.findOne({
-            $or: [{ username: identifier }, { email: identifier }]
-        });
-        return user;
-    } catch (error) {
-        throw new Error('User not found');
-    }
-};
-
-
-module.exports = {
-    getUser
-}
